pyproject.toml
[tool.poetry]
name = "nanobeepa"
version = "0.1.0"
description = "Backend for NanoBeepa"
authors = ["Kusuma <sensnerd@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.1"
djangorestframework = "^3.14.0"
django-cors-headers = "^4.0.0"
mysqlclient = "^2.1.1"
django-filter = "^23.2"
python-decouple = "^3.8"
djangorestframework-simplejwt = "^5.2.2"
cryptography = "^40.0.2"
pillow = "^9.5.0"
uuid = "^1.30"
django-allauth = "^0.54.0"
dj-rest-auth = "^4.0.1"
django-import-export = "^3.2.0"
midtransclient = "^1.4.1"
black = "^23.9.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
